requirements.txt
flask
pymongo
python-dotenv
pandas
tqdm
configparser
scikit-learn
